trace_analyzer/requirements.txt
google-cloud-aiplatform[adk,agent-engines]>=1.93.0
google-genai>=1.9.0
pydantic>=2.10.6
python-dotenv>=1.0.1
google-adk>=1.0.0
google-cloud-trace>=1.0.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp-proto-grpc>=1.24.0
google-auth>=2.18.1
grpcio>=1.63.0
numpy>=1.26.0
google-cloud-logging>=3.9.0
google-cloud-monitoring>=2.20.0
google-cloud-error-reporting>=1.14.0
